Switched to use directory_watcher gem
The inotify gem doesn't seem to want to build on my system, and
I also noticed it's aged a bit.  When I was installing some other
gem the other day, I noticed it depended on directory_watcher,
and this seems worth a try.  Instead of looking for close_write,
directory_watcher finds files added, modified, or removed.

Changed git add command to use -A, to note any changes, adds and
removals.

Signed-off-by: Chris Riddoch <riddochc@gmail.com>